+++ OnlineEducationPlatform.Web/Models/Question.cs
@@ -4,7 +4,7 @@
     {
         public int QuestionId { get; set; }
         public string Text { get; set; }
-        public decimal Points { get; set; }
+        public int Points { get; set; }
 
         // For multiple choice
         public List<string>? Options { get; set; } // Serialized list of choices
