+++ OnlineEducationPlatform.Web/ViewModels/QuestionViewModel.cs
@@ -6,7 +6,7 @@
         [System.ComponentModel.DataAnnotations.Required]
         public string Text { get; set; }
         [System.ComponentModel.DataAnnotations.Required]
-        public decimal Points { get; set; }
+        public int Points { get; set; }
         [System.ComponentModel.DataAnnotations.Required]
         public List<string>? Options { get; set; }
         [System.ComponentModel.DataAnnotations.Required]
